# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trançify SaaS multi-tenant scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/me:
    get:
      operationId: getMyTenant
      tags: [tenants]
      summary: Get current tenant profile
      responses:
        "200":
          description: Tenant info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMyTenant
      tags: [tenants]
      summary: Update current tenant profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/public/{slug}:
    get:
      operationId: getPublicTenant
      tags: [tenants]
      summary: Get public tenant info by slug (for client booking page)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public tenant info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services:
    get:
      operationId: getMyServices
      tags: [services]
      summary: Get services for current tenant
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /services/public/{tenantId}:
    get:
      operationId: getPublicServices
      tags: [services]
      summary: Get public services for a tenant (for client booking)
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Tenant:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        email:
          type: string
        whatsapp:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, blocked]
        plan:
          type: string
          enum: [basic, professional, enterprise]
        createdAt:
          type: string
          format: date-time
      required: [id, slug, name, email, status, plan, createdAt]
    PublicTenant:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        whatsapp:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
      required: [id, slug, name]
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
    Service:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        durationHours:
          type: number
        priceSmall:
          type: number
          description: Price for "up to mid-back" size
        priceLarge:
          type: number
          description: Price for "waist/butt" size
        sizeDependent:
          type: boolean
          description: Whether price varies by size
        active:
          type: boolean
      required: [id, tenantId, name, durationHours, priceSmall, priceLarge, sizeDependent, active]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationHours:
          type: number
        priceSmall:
          type: number
        priceLarge:
          type: number
        sizeDependent:
          type: boolean
      required: [name, durationHours, priceSmall, priceLarge, sizeDependent]
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationHours:
          type: number
        priceSmall:
          type: number
        priceLarge:
          type: number
        sizeDependent:
          type: boolean
        active:
          type: boolean
...